Fix dependency inversions before package split
- Extract STOP_WORDS/SYNONYMS from vrt/snapshot/expectation.ts into ai/nlp.ts
  so the ai layer no longer depends on snapshot
- Extract ViewportDetectionResult type from experiments/detection/detection-classify.ts
  into vrt/capture/detection-types.ts so the capture layer no longer depends
  on experiments; detection-classify now re-exports the type for callers
